pyproject.toml
[tool.poetry]
name = "fastapi-auth-base"
version = "0.1.0"
description = ""
authors = ["Firo <firocore@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
fastapi = "^0.114.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
loguru = "^0.7.2"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
